Enable full fastboot cap if it's not
Android Q has moved device-specific recovery extension to a
new shared library and the recovery executable dlopen() it to
make_device(). Unfortunately our extension demands access to
a variable, `reason`, which only exists in the executable. We
could not read that given we are a separate shared library, nor
can we call anything in the host executable.

However if this Android Q code is ever run, we can safely assume
that the device has been unlocked. Just proceed to set the full
cap flag, no need to check for the reason.

NOTE: recovery_reason has been removed - it does not seem proper
in this case, as it triggers another code path in coreboot.

Change-Id: I6aa0dfa6eaa1151be7d13cb1095ed240c009fb15